+++ CHANGELOG.md
@@ -14,6 +14,8 @@ a process;
 
 ### Changed
 
+- concord-runner: fix the handling of process arguments when
+restoring a process from a checkpoint;
 - concord-server, concord-agent: perform `git clone` on the Agent,
 keep only the changes in the DB;
 - bpm: updated to 0.51.0, fixed the resolving order of tasks and